fix(ui): move ui components out of literal @ dir and fix utils alias
shadcn CLI wrote button/input/sheet under a literal "@" folder and
pointed the utils alias at a nonexistent @/components/utils path,
breaking type resolution. Move the components under src/components/ui
and point aliases at the real @/lib/utils.
